Move DoDownload actions in its own class.
